# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blog API",
        "description": "Documentation de l’API du projet Blog",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Serveur principal"
        }
    ],
    "paths": {
...
        "/api/categories": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Lister toutes les catégories",
                "operationId": "0a3810ef08509327a6db442d3e4c4b73",
                "responses": {
                    "200": {
                        "description": "Liste des catégories récupérée avec succès"
                    }
                }
            },
            "post": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Créer une catégorie",
                "operationId": "f44066b315f1a6670d352e128ba39465",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nom"
                                ],
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Technologie"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Catégorie sur les articles tech"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Catégorie créée avec succès"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Afficher une catégorie spécifique",
                "description": "Récupère les détails d'une catégorie selon son ID",
                "operationId": "getCategorieById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant unique de la catégorie",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégorie trouvée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nom": {
                                            "type": "string",
                                            "example": "Technologie"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Articles sur les innovations tech"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-10-23T10:20:30"
                                        },
                                        "created_at_formatted": {
                                            "type": "string",
                                            "example": "2025-10-23 10:20:30"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catégorie introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Catégorie introuvable"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Mettre à jour une catégorie",
                "description": "Met à jour les informations d'une catégorie existante selon son ID.",
                "operationId": "updateCategorie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Identifiant unique de la catégorie à mettre à jour",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "description": "Les champs à mettre à jour pour la catégorie",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nom": {
                                        "type": "string",
                                        "example": "Technologie et Innovation"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Catégorie mise à jour avec des articles sur la tech moderne"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Catégorie mise à jour avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "nom": {
                                            "type": "string",
                                            "example": "Technologie et Innovation"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Catégorie mise à jour avec succès"
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-10-23T09:12:30"
                                        },
                                        "created_at_formatted": {
                                            "type": "string",
                                            "example": "2025-10-23 09:12:30"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catégorie introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Catégorie introuvable"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Catégories"
                ],
                "summary": "Supprimer une catégorie",
                "description": "Supprime une catégorie existante à partir de son ID.",
                "operationId": "deleteCategorie",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la catégorie à supprimer",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Catégorie supprimée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Catégorie supprimée avec succès"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Catégorie introuvable",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Catégorie introuvable"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}